backend/contact-form/requirements.txt
pydantic>=2.0.0
pydantic[email]>=2.0.0
psycopg2-binary>=2.9.0